+++ app/src/main/java/com/example/hundredplaces/data/model/visit/repositories/VisitsDataRepository.kt
@@ -17,7 +17,7 @@ class VisitsDataRepository(
         }
     }
 
-    override suspend fun getVisitStream(id: Int): Flow<Visit?> {
+    override suspend fun getVisitStream(id: Long): Flow<Visit?> {
         return if (networkConnection.isNetworkConnected) {
             visitsRemoteRepository.getVisitStream(id)
         } else {
